Move shared table code to classfile
