+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/impl/ModuleTable.java
@@ -30,7 +30,7 @@ import com.oracle.truffle.espresso.jdwp.api.ModuleRef;
 import com.oracle.truffle.espresso.meta.Meta;
 import com.oracle.truffle.espresso.runtime.staticobject.StaticObject;
 
-public final class ModuleTable extends com.oracle.truffle.espresso.impl.shared.ModuleTable<StaticObject, ModuleTable.ModuleEntry> {
+public final class ModuleTable extends com.oracle.truffle.espresso.classfile.tables.ModuleTable<StaticObject, ModuleTable.ModuleEntry> {
     public ModuleTable(ReadWriteLock lock) {
         super(lock);
     }
@@ -40,7 +40,7 @@ public final class ModuleTable extends com.oracle.truffle.espresso.impl.shared.M
         return new ModuleEntry(name, data);
     }
 
-    public static final class ModuleEntry extends com.oracle.truffle.espresso.impl.shared.ModuleTable.ModuleEntry<StaticObject> implements ModuleRef {
+    public static final class ModuleEntry extends com.oracle.truffle.espresso.classfile.tables.ModuleTable.ModuleEntry<StaticObject> implements ModuleRef {
         ModuleEntry(Symbol<Name> name, ModuleData<StaticObject> data) {
             super(name, data);
         }
